FDG-9170 create separate styling files
