+++ src/pages/unavailable/index.tsx
@@ -1,20 +1,19 @@
 import React from 'react';
+import { siteUnavailable, unavailableWrapper } from './unavailable.module.scss';
 import PageHelmet from '../../components/page-helmet/page-helmet';
 import SiteLayout from '../../components/siteLayout/siteLayout';
-import PageUnavailableText from '../../components/page-unavailable/page-unavilable-text';
+import PageUnavailableText from '../../components/page-unavailable/page-unavailable-text';
 
 const UnavailableContent = ({ fallback }) => {
   const pageTitle = fallback ? 'Content Currently Unavailable' : 'Site Outage Page';
 
   return (
-    <>
-      <div>
-        <PageHelmet data-testid="helmet" pageTitle={pageTitle} />
-        <div data-testid="unavailableWrapper">
-          <PageUnavailableText fallback={fallback} />
-        </div>
+    <div className={siteUnavailable}>
+      <PageHelmet data-testid="helmet" pageTitle={pageTitle} />
+      <div data-testid="unavailableWrapper" className={unavailableWrapper}>
+        <PageUnavailableText fallback={fallback} />
       </div>
-    </>
+    </div>
   );
 };
