+++ src/pages/unavailable/unavailable.module.scss
@@ -1,0 +1,35 @@
+@import '../../variables.module.scss';
+
+.siteUnavailable {
+  max-width: $main-width;
+  margin: 0 auto 1rem;
+  padding: 0 0.75rem;
+}
+.unavailableWrapper {
+  margin-bottom: 1rem;
+}
+
+@media screen and (min-width: $breakpoint-lg) and (max-width: 1204px) {
+  .unavailableWrapper {
+    display: flex;
+  }
+}
+
+@media screen and (min-width: $breakpoint-sm) {
+  .siteUnavailable {
+    padding: 1.25rem 2rem;
+  }
+}
+
+@media screen and (min-width: $breakpoint-xl) {
+  .siteUnavailable {
+    min-height: 34rem;
+    height: calc(100vw - 550px);
+    max-height: 48rem;
+  }
+
+  .unavailableWrapper {
+    display: flex;
+    align-items: stretch;
+  }
+}
